requirements.txt
Flask
numpy
joblib
pillow